restored README.MD and LICENSE files fixed global imports (reverted to importing individual classes) LocDups class split into two classes: LocDups contains the static part   and LocDupsApp is a regular class Digests class: method Combined refactored using streams,   LinkedHashMap used for storing digests values (as it preserves order)   ArrayList dropped   test included various method arguments and classes changed to final DupLocator class: corrected method getNamesakes()
