+++ src/main/java/org/lscode/DupLocator/FileArray.java
@@ -7,8 +7,11 @@
 
 package org.lscode.DupLocator;
 
-import java.io.File;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Set;
 
 public class FileArray implements Iterable<FileData>{
     private List<FileData> fileArray = new ArrayList<>();
