+++ src/main/java/org/lscode/DupLocator/FileStorage.java
@@ -7,7 +7,14 @@
 
 package org.lscode.DupLocator;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class FileStorage<T> implements Iterable<FileArray>{
